docs: update storage class recommendations for ROKS installs
Event Streams favours block storage, so switching the
 recommendation to use that instead.

EEM and EP are happy with file storage, however the file storage
classes (e.g. ibmc-file-bronze-gid, ibmc-file-silver-gid, etc.)
are not available in multizone clusters. We could leave the file
storage class in, and have a separate example for multi-zone
clusters, however I feel like that is unnecessarily complex - and
this example will work on both single-zone and multi-zone ROKS
clusters.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>