feat: add additional pre-reqs check to Event Streams install
The Event Streams install includes building a custom container
image (for Kafka Connect - the custom image has the connector used
for generating demo events).

This commit adds a check that this is going to be possible, with
a step that explains why it is needed if it fails on clusters
without an image registry.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>