fix: correct the scope of the Apicurio jars in the Connect runtime
Apicurio jars were in the Connect build as a plugin in their own
right, despite only being used by the loosehangerjeans-source
Connector.

This was risky, as it meant that if another Connector is added to
the Connect runtime that also happens to include a different
version of the io.apicurio.registry.utils.converter.AvroConverter
class, there is a chance that this will be used instead.

This commit modifies the Kaniko build config to more explicitly
scope the Apicurio classes as being used by the loosehangerjeans
Connector. This means that if additional Connectors are added to
the runtime that also have their version of AvroConverter, it will
not impact the datagen connector.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>