fix: update location of truststore for schema registry API calls (#44)
The Event Streams Operator used to make a truststore available in 
the Kafka Connect pod at a known location. The connector used in
the demo relied upon this by configuring the Apicurio converter to
reuse that truststore for REST API calls to the schema registry.

Event Streams operator behaviour has changed, and now doesn't have
that default mounted truststore. This commit fixes the connectors
that use the Apicurio converter by adding in a new explicit 
volume mount for the truststore file, and updating the converter
config to point at the new location.

Signed-off-by: A S Adil Mohammad <adil.mohammad@ibm.com>